dags/includes/requirements.txt
apache-airflow-providers-google==10.9.0
psycopg2-binary==2.9.9
apache-airflow-providers-celery==3.3.4
redis==5.0.1
pandas==2.1.1